refactor(repos,strategies): BaseSqliteRepository + PRICE_SCALE Konstante
Alle 8 SQLite-Repositories erben jetzt von BaseSqliteRepository (_base_sqlite.py),
das _get_connection(), _connection() und WAL/FK-Pragmas zentral definiert.
Entfernt ~120 Zeilen Duplikat-Boilerplate.

PRICE_SCALE (Decimal("0.0001")) aus rsi2.py und turtle.py in
strategies/_constants.py zentralisiert.
