+++ src/cilly_trading/repositories/analysis_runs_sqlite.py
@@ -6,41 +6,22 @@ from __future__ import annotations
 
 import json
 import sqlite3
-from contextlib import closing
 from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any, Dict, Optional
 
-from cilly_trading.db import DEFAULT_DB_PATH, init_db
 from cilly_trading.engine.core import AnalysisRun
+from cilly_trading.repositories._base_sqlite import BaseSqliteRepository
 from cilly_trading.repositories.analysis_run_evidence import (
     build_analysis_run_evidence_metadata,
     write_analysis_run_evidence_artifacts,
 )
 
 
-class SqliteAnalysisRunRepository:
+class SqliteAnalysisRunRepository(BaseSqliteRepository):
     """
     Repository für Analyse-Run-Metadaten und Ergebnisse.
     """
-
-    def __init__(self, db_path: Optional[Path] = None) -> None:
-        if db_path is None:
-            db_path = DEFAULT_DB_PATH
-
-        self._db_path = Path(db_path)
-        init_db(self._db_path)
-
-    def _get_connection(self) -> sqlite3.Connection:
-        conn = sqlite3.connect(self._db_path, timeout=5.0)
-        conn.row_factory = sqlite3.Row
-        conn.execute("PRAGMA journal_mode=WAL;")
-        conn.execute("PRAGMA foreign_keys = ON;")
-        conn.execute("PRAGMA busy_timeout = 5000;")
-        return conn
-
-    def _connection(self):
-        return closing(self._get_connection())
 
     @staticmethod
     def _deserialize_run_row(row: sqlite3.Row) -> Dict[str, Any]:
