+++ src/cilly_trading/repositories/execution_core_sqlite.py
@@ -3,33 +3,20 @@
 from __future__ import annotations
 
 import sqlite3
-from contextlib import closing
 from pathlib import Path
 from typing import Optional
 
-from cilly_trading.db import DEFAULT_DB_PATH, init_db
 from cilly_trading.models import ExecutionEvent, Order, Trade
 from cilly_trading.repositories import CanonicalExecutionRepository
+from cilly_trading.repositories._base_sqlite import BaseSqliteRepository
 
 
-class SqliteCanonicalExecutionRepository(CanonicalExecutionRepository):
+class SqliteCanonicalExecutionRepository(BaseSqliteRepository, CanonicalExecutionRepository):
     """Deterministic persistence for canonical orders, execution events, and trades."""
 
     def __init__(self, db_path: Optional[Path] = None) -> None:
-        self._db_path = Path(db_path or DEFAULT_DB_PATH)
-        init_db(self._db_path)
+        super().__init__(db_path)
         self._ensure_tables()
-
-    def _get_connection(self) -> sqlite3.Connection:
-        conn = sqlite3.connect(self._db_path, timeout=5.0)
-        conn.row_factory = sqlite3.Row
-        conn.execute("PRAGMA journal_mode=WAL;")
-        conn.execute("PRAGMA foreign_keys = ON;")
-        conn.execute("PRAGMA busy_timeout = 5000;")
-        return conn
-
-    def _connection(self):
-        return closing(self._get_connection())
 
     def _ensure_tables(self) -> None:
         with self._connection() as conn:
