+++ src/cilly_trading/repositories/snapshot_ingestion_sqlite.py
@@ -2,12 +2,11 @@ from __future__ import annotations
 
 import json
 import sqlite3
-from contextlib import closing
 from dataclasses import dataclass
 from pathlib import Path
 from typing import Optional
 
-from cilly_trading.db import DEFAULT_DB_PATH, init_db
+from cilly_trading.repositories._base_sqlite import BaseSqliteRepository
 
 
 @dataclass(frozen=True)
@@ -33,25 +32,8 @@ class OhlcvSnapshotRowRecord:
     volume: float
 
 
-class SqliteSnapshotIngestionRepository:
+class SqliteSnapshotIngestionRepository(BaseSqliteRepository):
     """Persist bounded snapshot ingestion runs into the canonical SQLite schema."""
-
-    def __init__(self, db_path: Optional[Path] = None) -> None:
-        if db_path is None:
-            db_path = DEFAULT_DB_PATH
-        self._db_path = Path(db_path)
-        init_db(self._db_path)
-
-    def _get_connection(self) -> sqlite3.Connection:
-        conn = sqlite3.connect(self._db_path, timeout=5.0)
-        conn.row_factory = sqlite3.Row
-        conn.execute("PRAGMA journal_mode=WAL;")
-        conn.execute("PRAGMA foreign_keys = ON;")
-        conn.execute("PRAGMA busy_timeout = 5000;")
-        return conn
-
-    def _connection(self):
-        return closing(self._get_connection())
 
     def save_snapshot_run(
         self,
