+++ src/cilly_trading/repositories/trades_sqlite.py
@@ -3,35 +3,16 @@
 from __future__ import annotations
 
 import sqlite3
-from contextlib import closing
 from pathlib import Path
 from typing import List, Optional
 
-from cilly_trading.db import init_db, DEFAULT_DB_PATH  # type: ignore
 from cilly_trading.models import PersistedTradePayload
 from cilly_trading.repositories import TradeRepository
+from cilly_trading.repositories._base_sqlite import BaseSqliteRepository
 
 
-class SqliteTradeRepository(TradeRepository):
+class SqliteTradeRepository(BaseSqliteRepository, TradeRepository):
     """Speichert und lädt Trades aus einer SQLite-Datenbank."""
-
-    def __init__(self, db_path: Optional[Path] = None) -> None:
-        if db_path is None:
-            db_path = DEFAULT_DB_PATH
-
-        self._db_path = Path(db_path)
-        init_db(self._db_path)
-
-    def _get_connection(self) -> sqlite3.Connection:
-        conn = sqlite3.connect(self._db_path, timeout=5.0)
-        conn.row_factory = sqlite3.Row
-        conn.execute("PRAGMA journal_mode=WAL;")
-        conn.execute("PRAGMA foreign_keys = ON;")
-        conn.execute("PRAGMA busy_timeout = 5000;")
-        return conn
-
-    def _connection(self):
-        return closing(self._get_connection())
 
     def save_trade(self, trade: PersistedTradePayload) -> int:
         with self._connection() as conn:
