+++ src/cilly_trading/strategies/_constants.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from decimal import Decimal
+
+PRICE_SCALE: Decimal = Decimal("0.0001")
