+++ src/cilly_trading/strategies/turtle.py
@@ -24,6 +24,7 @@ import pandas as pd
 
 from cilly_trading.models import Signal
 from cilly_trading.engine.core import BaseStrategy
+from cilly_trading.strategies._constants import PRICE_SCALE
 
 
 @dataclass
@@ -120,12 +121,12 @@ class TurtleStrategy(BaseStrategy):
                 "Breakout-Hoch oder unter einem definierten Trailing-Stop)."
             )
 
-            _scale = Decimal("0.0001")
+
             _stop = float(
                 (
                     Decimal(str(breakout_level))
                     * (Decimal("1") - Decimal(str(cfg.stop_loss_buffer_pct)))
-                ).quantize(_scale, ROUND_HALF_UP)
+                ).quantize(PRICE_SCALE, ROUND_HALF_UP)
             )
             signal: Signal = {
                 "strategy": self.name,
@@ -135,10 +136,10 @@ class TurtleStrategy(BaseStrategy):
                 "confirmation_rule": confirmation_rule,
                 "entry_zone": {
                     "from_": float(
-                        Decimal(str(breakout_level)).quantize(_scale, ROUND_HALF_UP)
+                        Decimal(str(breakout_level)).quantize(PRICE_SCALE, ROUND_HALF_UP)
                     ),
                     "to": float(
-                        (Decimal(str(last_close)) * Decimal("1.02")).quantize(_scale, ROUND_HALF_UP)
+                        (Decimal(str(last_close)) * Decimal("1.02")).quantize(PRICE_SCALE, ROUND_HALF_UP)
                     ),
                 },
                 "stop_loss": _stop,
@@ -163,12 +164,12 @@ class TurtleStrategy(BaseStrategy):
                         "Alternativ: Stop-Buy-Order knapp über dem Breakout-Hoch."
                     )
 
-                    _scale = Decimal("0.0001")
+        
                     _stop = float(
                         (
                             Decimal(str(breakout_level))
                             * (Decimal("1") - Decimal(str(cfg.stop_loss_buffer_pct)))
-                        ).quantize(_scale, ROUND_HALF_UP)
+                        ).quantize(PRICE_SCALE, ROUND_HALF_UP)
                     )
                     signal = {
                         "strategy": self.name,
@@ -181,11 +182,11 @@ class TurtleStrategy(BaseStrategy):
                                 (
                                     Decimal(str(breakout_level))
                                     * (Decimal("1") - Decimal(str(cfg.proximity_threshold_pct)))
-                                ).quantize(_scale, ROUND_HALF_UP)
+                                ).quantize(PRICE_SCALE, ROUND_HALF_UP)
                             ),
                             "to": float(
                                 (Decimal(str(breakout_level)) * Decimal("1.01")).quantize(
-                                    _scale, ROUND_HALF_UP
+                                    PRICE_SCALE, ROUND_HALF_UP
                                 )
                             ),
                         },
